Updated components, added some presentational components
