+++ src/reactjs/App.js
@@ -17,45 +17,12 @@ class App extends React.Component {
         this.props.fetchData();
     }
 
-    shouldRenderPromoList (props) {
-        const { search } = props.location;
-        if (search && search.indexOf("?") !== -1) {
-            return false;
-        }
-        return true;
-    }
-
-    getRequestedPromoIndex (props) {
-        const { search } = props.location;
-        const params = new URLSearchParams(search);
-        const promoID = params.get('promo');
-        if (promoID === null) {
-            return null;
-        }
-
-        const promoIdRet = promoID.replace(/promo/g,'');
-        const promoIndex = parseInt(promoIdRet, 0) - 1;
-        return promoIndex;
-    }
-
-    renderViews (props) {
-        if (this.shouldRenderPromoList(props) === true) {
-            return <PromotionListView />;
-        } else {
-            const requestedPromoIndex = this.getRequestedPromoIndex(props);
-            if (requestedPromoIndex !== null) {
-                return <PromotionView promotionIndex={requestedPromoIndex} />;
-            } else {
-                return <NotFoundPage />;
-            }
-        }
-    }
-
 	render() {
 		return (
             <BrowserRouter>
                 <Switch>
-                    <Route path="/index.html" render={(props) => this.renderViews(props)} />
+                    <Route path="/" component={PromotionListView} exact />
+                    <Route path="/promo/:id" component={PromotionView} />
                     <Route component={NotFoundPage} />
                 </Switch>
             </BrowserRouter>
