+++ src/reactjs/components/PromotionList.js
@@ -2,30 +2,28 @@ import React from "react";
 import numeral from "numeral";
 import PromotionListItem from "./PromotionListItem";
 
-class PromotionList extends React.Component {
-    render () {
-        const { promotion_objects: promotionsData, server_time: serverTimeData } = this.props.data;
-        const promotionListItems = promotionsData.map((promotionData, index) => {
-            const key = `promotionListItem${index}`;
-            const url = `?promo=promo${numeral(index + 1).format("00")}`;
-            return (
-                <PromotionListItem
-                    key={key}
-                    url={url}
-                    serverTimeData={serverTimeData}
-                    promotionData={promotionData}
-                />
-            );
-        });
-
+const PromotionList = ({ data }) => {
+    const { promotion_objects: promotionsData, server_time: serverTimeData } = data;
+    const promotionListItems = promotionsData.map((promotionData, index) => {
+        const key = `promotionListItem${index}`;
+        const url = `/promo/promo${numeral(index + 1).format("00")}`;
         return (
-            <div className="promotionListView">
-                <div className="promotionList container">
-                    { promotionListItems }
-                </div>
+            <PromotionListItem
+                key={key}
+                url={url}
+                serverTimeData={serverTimeData}
+                promotionData={promotionData}
+            />
+        );
+    });
+
+    return (
+        <div className="promotionListView">
+            <div className="promotionList container">
+                { promotionListItems }
             </div>
-        )
-    }
-}
+        </div>
+    );
+};
 
 export default PromotionList;
